test/requirements.txt
# melonDS DS doesn't support OpenGL on macOS,
# so there's no need to install OpenGL dependencies
libretro.py==0.3.0
libretro.py[opengl]==0.3.0 ; sys_platform != "darwin"
Pillow==10.3.0
more-itertools==10.2.* ; python_version < '3.12'
